Rearrange test files and touch up
